+++ ErsatzTV.Application/Channels/Commands/UpdateChannel.cs
@@ -14,14 +14,5 @@ namespace ErsatzTV.Application.Channels.Commands
         string Logo,
         string PreferredLanguageCode,
         StreamingMode StreamingMode,
-        string Watermark,
-        ChannelWatermarkMode WatermarkMode,
-        ChannelWatermarkLocation WatermarkLocation,
-        ChannelWatermarkSize WatermarkSize,
-        int WatermarkWidth,
-        int WatermarkHorizontalMargin,
-        int WatermarkVerticalMargin,
-        int WatermarkFrequencyMinutes,
-        int WatermarkDurationSeconds,
-        int WatermarkOpacity) : IRequest<Either<BaseError, ChannelViewModel>>;
+        int? WatermarkId) : IRequest<Either<BaseError, ChannelViewModel>>;
 }