+++ ErsatzTV.Application/Watermarks/Queries/GetWatermarkById.cs
@@ -1,0 +1,7 @@
+﻿using LanguageExt;
+using MediatR;
+
+namespace ErsatzTV.Application.Watermarks.Queries
+{
+    public record GetWatermarkById(int Id) : IRequest<Option<WatermarkViewModel>>;
+}