+++ ErsatzTV.Infrastructure/Data/TvContext.cs
@@ -14,6 +14,7 @@ namespace ErsatzTV.Infrastructure.Data
 
         public DbSet<ConfigElement> ConfigElements { get; set; }
         public DbSet<Channel> Channels { get; set; }
+        public DbSet<ChannelWatermark> ChannelWatermarks { get; set; }
         public DbSet<MediaSource> MediaSources { get; set; }
         public DbSet<LocalMediaSource> LocalMediaSources { get; set; }
         public DbSet<PlexMediaSource> PlexMediaSources { get; set; }